+++ src/dawn_native/metal/DeviceMTL.h
@@ -56,7 +56,8 @@ namespace dawn_native { namespace metal {
                                                     uint32_t plane);
         void WaitForCommandsToBeScheduled();
 
-        ResultOrError<std::unique_ptr<StagingBufferBase>> CreateStagingBuffer(size_t size) override;
+        ResultOrError<std::unique_ptr<StagingBufferBase>> CreateStagingBuffer(
+            uint64_t size) override;
         MaybeError CopyFromStagingToBuffer(StagingBufferBase* source,
                                            uint64_t sourceOffset,
                                            BufferBase* destination,
