+++ src/dawn_native/vulkan/DeviceVk.h
@@ -70,7 +70,8 @@ namespace dawn_native { namespace vulkan {
         Serial GetLastSubmittedCommandSerial() const final override;
         void TickImpl() override;
 
-        ResultOrError<std::unique_ptr<StagingBufferBase>> CreateStagingBuffer(size_t size) override;
+        ResultOrError<std::unique_ptr<StagingBufferBase>> CreateStagingBuffer(
+            uint64_t size) override;
         MaybeError CopyFromStagingToBuffer(StagingBufferBase* source,
                                            uint64_t sourceOffset,
                                            BufferBase* destination,
